Move non-application-based components into misc folder
