+++ misc/mixins/graph.py
